Added support of text styles in messages

+++ Sources/Falko.Talkie.Bridges.Telegram/Models/Update.cs
@@ -1,7 +1,7 @@
 namespace Talkie.Bridges.Telegram.Models;
 
 public sealed class Update(
-    long updateId,
+    int updateId,
     Message? message = null,
     Message? editedMessage = null,
     Message? channelPost = null,
@@ -9,7 +9,7 @@ public sealed class Update(
     Message? businessMessage = null,
     Message? editedBusinessMessage = null)
 {
-    public readonly long UpdateId = updateId;
+    public readonly int UpdateId = updateId;
 
     public readonly Message? Message = message;
